ESQL: Add interfaces to distribute the post-analysis verification (#119798)
This adds a PostAnalysisVerificationAware interface that allows an expression,
plan or even command to perform post-analysis verifications "locally", vs.
having them centralized in the core verifier.

(cherry picked from commit ad264f7bf392a563688b94744e630dfbc7c4aef6)
