+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/function/grouping/GroupingFunction.java
@@ -7,14 +7,21 @@
 
 package org.elasticsearch.xpack.esql.expression.function.grouping;
 
+import org.elasticsearch.xpack.esql.capabilities.PostAnalysisPlanVerificationAware;
+import org.elasticsearch.xpack.esql.common.Failures;
 import org.elasticsearch.xpack.esql.core.expression.Expression;
 import org.elasticsearch.xpack.esql.core.expression.function.Function;
 import org.elasticsearch.xpack.esql.core.tree.Source;
 import org.elasticsearch.xpack.esql.evaluator.mapper.EvaluatorMapper;
+import org.elasticsearch.xpack.esql.plan.logical.Aggregate;
+import org.elasticsearch.xpack.esql.plan.logical.LogicalPlan;
 
 import java.util.List;
+import java.util.function.BiConsumer;
 
-public abstract class GroupingFunction extends Function implements EvaluatorMapper {
+import static org.elasticsearch.xpack.esql.common.Failure.fail;
+
+public abstract class GroupingFunction extends Function implements EvaluatorMapper, PostAnalysisPlanVerificationAware {
 
     protected GroupingFunction(Source source, List<Expression> fields) {
         super(source, fields);
@@ -23,6 +30,18 @@ public abstract class GroupingFunction extends Function implements EvaluatorMapp
     @Override
     public Object fold() {
         return EvaluatorMapper.super.fold();
+    }
+
+    @Override
+    public BiConsumer<LogicalPlan, Failures> postAnalysisPlanVerification() {
+        return (p, failures) -> {
+            if (p instanceof Aggregate == false) {
+                p.forEachExpression(
+                    GroupingFunction.class,
+                    gf -> failures.add(fail(gf, "cannot use grouping function [{}] outside of a STATS command", gf.sourceText()))
+                );
+            }
+        };
     }
 
 }
