+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/plan/logical/Filter.java
@@ -9,6 +9,8 @@ package org.elasticsearch.xpack.esql.plan.logical;
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.xpack.esql.capabilities.PostAnalysisVerificationAware;
+import org.elasticsearch.xpack.esql.common.Failures;
 import org.elasticsearch.xpack.esql.core.expression.Expression;
 import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
 import org.elasticsearch.xpack.esql.core.tree.Source;
@@ -17,12 +19,16 @@ import org.elasticsearch.xpack.esql.io.stream.PlanStreamInput;
 import java.io.IOException;
 import java.util.Objects;
 
+import static org.elasticsearch.xpack.esql.common.Failure.fail;
+import static org.elasticsearch.xpack.esql.core.type.DataType.BOOLEAN;
+import static org.elasticsearch.xpack.esql.core.type.DataType.NULL;
+
 /**
  * A {@code Filter} is a type of Plan that performs filtering of results. In
  * {@code SELECT x FROM y WHERE z ..} the "WHERE" clause is a Filter. A
  * {@code Filter} has a "condition" Expression that does the filtering.
  */
-public class Filter extends UnaryPlan {
+public class Filter extends UnaryPlan implements PostAnalysisVerificationAware {
     public static final NamedWriteableRegistry.Entry ENTRY = new NamedWriteableRegistry.Entry(LogicalPlan.class, "Filter", Filter::new);
 
     private final Expression condition;
@@ -97,5 +103,16 @@ public class Filter extends UnaryPlan {
 
     public Filter with(LogicalPlan child, Expression conditionExpr) {
         return new Filter(source(), child, conditionExpr);
+    }
+
+    @Override
+    public void postAnalysisVerification(Failures failures) {
+        checkFilterConditionDataType(condition, failures);
+    }
+
+    public static void checkFilterConditionDataType(Expression expression, Failures failures) {
+        if (expression.dataType() != NULL && expression.dataType() != BOOLEAN) {
+            failures.add(fail(expression, "Condition expression needs to be boolean, found [{}]", expression.dataType()));
+        }
     }
 }
