+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/plan/logical/OrderBy.java
@@ -9,9 +9,12 @@ package org.elasticsearch.xpack.esql.plan.logical;
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.xpack.esql.capabilities.PostAnalysisVerificationAware;
+import org.elasticsearch.xpack.esql.common.Failures;
 import org.elasticsearch.xpack.esql.core.capabilities.Resolvables;
 import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
 import org.elasticsearch.xpack.esql.core.tree.Source;
+import org.elasticsearch.xpack.esql.core.type.DataType;
 import org.elasticsearch.xpack.esql.expression.Order;
 import org.elasticsearch.xpack.esql.io.stream.PlanStreamInput;
 
@@ -19,7 +22,9 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Objects;
 
-public class OrderBy extends UnaryPlan {
+import static org.elasticsearch.xpack.esql.common.Failure.fail;
+
+public class OrderBy extends UnaryPlan implements PostAnalysisVerificationAware {
     public static final NamedWriteableRegistry.Entry ENTRY = new NamedWriteableRegistry.Entry(LogicalPlan.class, "OrderBy", OrderBy::new);
 
     private final List<Order> order;
@@ -90,5 +95,17 @@ public class OrderBy extends UnaryPlan {
 
         OrderBy other = (OrderBy) obj;
         return Objects.equals(order, other.order) && Objects.equals(child(), other.child());
+    }
+
+    @Override
+    public void postAnalysisVerification(Failures failures) {
+        /**
+         * Some datatypes are not sortable
+         */
+        order.forEach(order -> {
+            if (DataType.isSortable(order.dataType()) == false) {
+                failures.add(fail(order, "cannot sort on " + order.dataType().typeName()));
+            }
+        });
     }
 }
