+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/plan/logical/Row.java
@@ -8,17 +8,22 @@
 package org.elasticsearch.xpack.esql.plan.logical;
 
 import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.xpack.esql.capabilities.PostAnalysisVerificationAware;
+import org.elasticsearch.xpack.esql.common.Failures;
 import org.elasticsearch.xpack.esql.core.capabilities.Resolvables;
 import org.elasticsearch.xpack.esql.core.expression.Alias;
 import org.elasticsearch.xpack.esql.core.expression.Attribute;
 import org.elasticsearch.xpack.esql.core.expression.Expressions;
 import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
 import org.elasticsearch.xpack.esql.core.tree.Source;
+import org.elasticsearch.xpack.esql.core.type.DataType;
 
 import java.util.List;
 import java.util.Objects;
 
-public class Row extends LeafPlan {
+import static org.elasticsearch.xpack.esql.common.Failure.fail;
+
+public class Row extends LeafPlan implements PostAnalysisVerificationAware {
 
     private final List<Alias> fields;
 
@@ -72,5 +77,14 @@ public class Row extends LeafPlan {
     @Override
     public int hashCode() {
         return Objects.hash(fields);
+    }
+
+    @Override
+    public void postAnalysisVerification(Failures failures) {
+        fields.forEach(a -> {
+            if (DataType.isRepresentable(a.dataType()) == false) {
+                failures.add(fail(a.child(), "cannot use [{}] directly in a row assignment", a.child().sourceText()));
+            }
+        });
     }
 }
